Use long long (64bit) for entry numbers in progress monitor
This is the type ROOT uses for this, corresponds to Int_64.

+++ RootTools/FileInterfaceBase.h
@@ -33,7 +33,7 @@ public:
 	{
 		return eventchain->GetEntries();
 	}
-	inline void GetEntry(size_t entry)
+	inline void GetEntry(long long entry)
 	{
 		eventchain->GetEntry(entry);
 	}
